+++ internal/app/helpers/htmx.go
@@ -1,0 +1,7 @@
+package helpers
+
+import "net/http"
+
+func IsHtmxRequest(req *http.Request) bool {
+	return req.Header.Get("HX-Request") == "true"
+}