+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 History of **Houston we have (no) problem** application with changes description.
 
 ## Next release
+
+- `FEATURE` Initialize application (GoLang, SQLite, Templ, TailwindCSS, FlyonUI, HTMX, Air)